+++ src/main/java/domain/deck/Symbol.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.deck;
 
 public enum Symbol {
