+++ src/main/java/domain/user/Dealer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.user;
 
 public class Dealer extends User {
