+++ src/test/java/domain/deck/CardTest.java
@@ -1,9 +1,13 @@
-package domain;
+package domain.deck;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import domain.deck.Card;
+import domain.deck.Symbol;
+import domain.deck.Type;
 
 class CardTest {
