+++ src/test/java/domain/deck/DeckFactoryTest.java
@@ -1,9 +1,11 @@
-package domain;
+package domain.deck;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import domain.deck.DeckFactory;
 
 class DeckFactoryTest {
