+++ src/test/java/domain/deck/SymbolTest.java
@@ -1,8 +1,10 @@
-package domain;
+package domain.deck;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+
+import domain.deck.Symbol;
 
 class SymbolTest {
