+++ src/test/java/domain/user/DealerTest.java
@@ -1,8 +1,10 @@
-package domain;
+package domain.user;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+
+import domain.user.Dealer;
 
 class DealerTest {
