+++ src/test/java/domain/user/PlayerTest.java
@@ -1,9 +1,11 @@
-package domain;
+package domain.user;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import domain.user.Player;
 
 class PlayerTest {
